Fix delta metric baselines for multiple readers
