+++ .changeset/fresh-rivers-report.md
@@ -1,0 +1,5 @@
+---
+"@effect/opentelemetry": patch
+---
+
+Isolate delta metric baselines for each registered metric reader
