+++ packages/opentelemetry/test/OtelMetrics.test.ts
@@ -1,7 +1,9 @@
 import * as internal from "@effect/opentelemetry/internal/metrics"
+import * as OtelMetrics from "@effect/opentelemetry/OtelMetrics"
 import { assert, describe, it } from "@effect/vitest"
 import { ValueType } from "@opentelemetry/api"
 import { resourceFromAttributes } from "@opentelemetry/resources"
+import { MetricReader } from "@opentelemetry/sdk-metrics"
 import * as Effect from "effect/Effect"
 import * as Metric from "effect/Metric"
 
@@ -336,4 +338,29 @@ describe("Metrics", () => {
         ]
       })
     }))
+
+  it.effect("reports the same delta to every registered reader", () =>
+    Effect.scoped(Effect.gen(function*() {
+      class Reader extends MetricReader {
+        protected onShutdown(): Promise<void> {
+          return Promise.resolve()
+        }
+        protected onForceFlush(): Promise<void> {
+          return Promise.resolve()
+        }
+      }
+
+      const services = yield* Effect.context<never>()
+      const producer = new internal.MetricProducerImpl(resourceFromAttributes({}), services, "delta")
+      const first = new Reader()
+      const second = new Reader()
+      yield* OtelMetrics.registerProducer(producer, () => [first, second])
+      yield* Metric.update(Metric.counter("requests", { incremental: true }), 1)
+
+      const firstResult = yield* Effect.promise(() => first.collect())
+      const secondResult = yield* Effect.promise(() => second.collect())
+      const firstValue = (firstResult.resourceMetrics.scopeMetrics[0]!.metrics[0] as any).dataPoints[0].value
+      const secondValue = (secondResult.resourceMetrics.scopeMetrics[0]!.metrics[0] as any).dataPoints[0].value
+      assert.deepStrictEqual([firstValue, secondValue], [1, 1])
+    })).pipe(Effect.provideService(Metric.MetricRegistry, new Map())))
 })
